Export macro assert_files_equal
